@@ -3,38 +3,38 @@
     "version": "0.0.1",
     "private": true,
     "dependencies": {
-        "@storybook/preset-create-react-app": "^3.2.0",
-        "@storybook/react": "^6.3.6",
-        "@testing-library/jest-dom": "^5.11.4",
-        "@testing-library/react": "^11.1.0",
-        "@testing-library/user-event": "^12.1.10",
-        "@types/faker": "^5.5.7",
-        "@types/jest": "^26.0.15",
-        "@types/lodash": "^4.14.182",
-        "@types/node": "^12.0.0",
-        "@types/react": "^17.0.0",
-        "@types/react-dom": "^17.0.0",
-        "@types/react-router-dom": "^5.1.8",
-        "@types/styled-components": "^5.1.11",
-        "@types/styled-theming": "^2.2.5",
-        "@xstate/react": "^1.5.1",
+        "@storybook/preset-create-react-app": "^4.1.2",
+        "@storybook/react": "^6.5.12",
+        "@testing-library/jest-dom": "^5.16.5",
+        "@testing-library/react": "^13.4.0",
+        "@testing-library/user-event": "^14.4.3",
+        "@types/jest": "^29.2.0",
+        "@types/lodash": "^4.14.186",
+        "@types/node": "^18.11.3",
+        "@xstate/react": "^3.0.1",
         "faker": "^5.5.3",
         "lodash": "^4.17.21",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "react-router-dom": "^5.2.0",
-        "react-scripts": "4.0.3",
-        "socket.io-client": "^4.1.3",
-        "styled-components": "^5.3.0",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-router-dom": "^6.4.2",
+        "react-scripts": "5.0.1",
+        "socket.io-client": "^4.5.3",
+        "styled-components": "^5.3.6",
         "styled-theming": "^2.2.0",
-        "typescript": "^4.1.2",
-        "xstate": "^4.23.1"
+        "typescript": "^4.8.4",
+        "xstate": "^4.33.6"
     },
     "devDependencies": {
-        "@storybook/addon-actions": "^6.3.6",
-        "@storybook/addon-essentials": "^6.3.6",
-        "@storybook/addon-links": "^6.3.6",
-        "@storybook/node-logger": "^6.3.6"
+        "@storybook/addon-actions": "^6.5.12",
+        "@storybook/addon-essentials": "^6.5.12",
+        "@storybook/addon-links": "^6.5.12",
+        "@storybook/node-logger": "^6.5.12",
+        "@types/faker": "^5.5.9",
+        "@types/react": "^18.0.21",
+        "@types/react-dom": "^18.0.6",
+        "@types/react-router-dom": "^5.3.3",
+        "@types/styled-components": "^5.1.26",
+        "@types/styled-theming": "^2.2.5"
     },
     "resolutions": {
         "styled-components": "^5"
