@@ -1,4 +1,4 @@
-import { Route, Switch } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 
 import { AppComponent as App } from './app.component';
 import { ROUTES } from './app.constants';
@@ -6,17 +6,17 @@ import { NotFound } from './notFound';
 import { Home } from './home';
 
 export const Main = () => {
-  return (
-    <App>
-      <Switch>
-        <Route exact path={ROUTES.home}>
-          <Home />
-        </Route>
+    return (
+        <App>
+            <Routes>
+                <Route path={ROUTES.home}>
+                    <Home />
+                </Route>
 
-        <Route>
-          <NotFound />
-        </Route>
-      </Switch>
-    </App>
-  );
+                <Route>
+                    <NotFound />
+                </Route>
+            </Routes>
+        </App>
+    );
 };
