@@ -1,19 +1,18 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Router } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 
-import browserHistory from './utils/history';
 import { Main } from './routes';
 
 const render = (): void => {
-  ReactDOM.render(
-    <React.StrictMode>
-      <Router history={browserHistory}>
-        <Main />
-      </Router>
-    </React.StrictMode>,
-    document.getElementById('root')
-  );
+    ReactDOM.render(
+        <React.StrictMode>
+            <BrowserRouter>
+                <Main />
+            </BrowserRouter>
+        </React.StrictMode>,
+        document.getElementById('root'),
+    );
 };
 
 render();
