@@ -1,16 +1,12 @@
 import React from 'react';
 
-import { Container, Place, Points } from './score.styles';
+import { Container } from './score.styles';
 
 interface ScoreProps {
-  place?: number;
-  points?: number;
+    place?: number;
+    points?: number;
 }
 
 export const Score = ({ place = 0, points = 0 }: ScoreProps) => {
-  return (
-    <Container>
-      Thank you for participating
-    </Container>
-  );
-}
+    return <Container>Thank you for participating</Container>;
+};
